requirements.txt
dnspython==2.6.1
Faker==28.0.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
